requirements.txt
fastmcp>=0.1.0
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0
